+++ tests/test.js
@@ -2,7 +2,7 @@
 *  Contemplate
 *  Light-weight Templating Engine for PHP, Python, Node and client-side JavaScript
 *
-*  @version 0.6.12
+*  @version 0.7
 *  https://github.com/foo123/Contemplate
 *
 *  @inspired by : Simple JavaScript Templating, John Resig - http://ejohn.org/ - MIT Licensed
@@ -42,7 +42,7 @@ global.bracket = function(v)
 {
     return '[[' + v + ']]';
 }
-Contemplate.addPlugin('inlinedBracket', null, 'bracket');
+Contemplate.addPlugin('inlinedBracket', Contemplate.inline('bracket'));
 
 // make sure it exists
 Contemplate.setCacheDir(fs.realpathSync(path.join(__dirname, '/_tplcache')));