+++ tests/test.php
@@ -3,7 +3,7 @@
 *  Contemplate
 *  Light-weight Templating Engine for PHP, Python, Node and client-side JavaScript
 *
-*  @version 0.6.12
+*  @version 0.7
 *  https://github.com/foo123/Contemplate
 *
 *  @inspired by : Simple JavaScript Templating, John Resig - http://ejohn.org/ - MIT Licensed
@@ -45,7 +45,7 @@ function bracket($v=null)
 {
     return '[[' . $v . ']]';
 }
-Contemplate::addPlugin('inlinedBracket', null, 'bracket');
+Contemplate::addPlugin('inlinedBracket', Contemplate::inline('bracket'));
 
 // set the cache directory (make sure to exist)
 Contemplate::setCacheDir(ABSPATH.'/_tplcache');