+++ changelog.md
@@ -1,5 +1,10 @@
 ###ChangeLog
 
+__0.7__
+* inline templates both in Contemplate templates (e.g as parameters to other functions) or in code
+* %inline template function, creates/renders inline templates
+* minor updates
+
 __0.6.12__
 * python-style for-loops contructs (see manual and examples)